+++ src/components/icon.rs
@@ -577,3 +577,33 @@ pub fn IconSquare2x2(
         </SvgWrapper>
     }
 }
+
+#[component]
+pub fn IconArrowUTurnLeft(
+    #[prop(optional)] size: Option<usize>,
+    #[prop(attrs)] attrs: Vec<(&'static str, Attribute)>,
+) -> impl IntoView {
+    view! {
+        <SvgWrapper size attrs>
+            <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                d="M9 15 3 9m0 0 6-6M3 9h12a6 6 0 0 1 0 12h-3"
+            ></path>
+        </SvgWrapper>
+    }
+}
+
+#[component]
+pub fn IconThreeDots(
+    #[prop(optional)] size: Option<usize>,
+    #[prop(attrs)] attrs: Vec<(&'static str, Attribute)>,
+) -> impl IntoView {
+    view! {
+        <SvgWrapper size attrs>
+            <circle cx="12" cy="12" r="1"></circle>
+            <circle cx="19" cy="12" r="1"></circle>
+            <circle cx="5" cy="12" r="1"></circle>
+        </SvgWrapper>
+    }
+}